nfsd4: keep a reference count on client while in use
Patchwork-id: 25883
O-Subject: [RHEL6 PATCH 39/47] nfsd4: keep a reference count on client while in
	use
Bugzilla: 597215
RH-Acked-by: Jeff Layton <jlayton@redhat.com>

From: Benny Halevy <bhalevy@panasas.com>

Get a refcount on the client on SEQUENCE,
Release the refcount and renew the client when all respective compounds completed.
Do not expire the client by the laundromat while in use.
If the client was expired via another path, free it when the compounds
complete and the refcount reaches 0.

Note that unhash_client_locked must call list_del_init on cl_lru as
it may be called twice for the same client (once from nfs4_laundromat
and then from expire_client)

Signed-off-by: Benny Halevy <bhalevy@panasas.com>
Signed-off-by: J. Bruce Fields <bfields@citi.umich.edu>
Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>