Revert "nfsd4: distinguish expired from stale stateids"
Patchwork-id: 25887
O-Subject: [RHEL6 PATCH 45/47] Revert "nfsd4: distinguish expired from stale
	stateids"
Bugzilla: 597215
RH-Acked-by: Jeff Layton <jlayton@redhat.com>

From: J. Bruce Fields <bfields@citi.umich.edu>

This reverts commit 78155ed75f470710f2aecb3e75e3d97107ba8374.

We're depending here on the boot time that we use to generate the
stateid being monotonic, but get_seconds() is not necessarily.

We still depend at least on boot_time being different every time, but
that is a safer bet.

We have a few reports of errors that might be explained by this problem,
though we haven't been able to confirm any of them.

But the minor gain of distinguishing expired from stale errors seems not
worth the risk.

Conflicts:

	fs/nfsd/nfs4state.c

Signed-off-by: J. Bruce Fields <bfields@citi.umich.edu>
Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>