virtio: console: Fix crash when port is unplugged and blocked for write
Patchwork-id: 25894
O-Subject: [RHEL6 PATCH 2/2] virtio: console: Fix crash when port is unplugged
	and blocked for write
Bugzilla: 596635
RH-Acked-by: Jes Sorensen <Jes.Sorensen@redhat.com>

When a program that has a virtio port opened and blocked for a write
operation, a port hot-unplug event will later led to a crash when
SIGTERM was sent to the program. Fix that.

Bugzilla: 596635
Upstream status: posted

Signed-off-by: Amit Shah <amit.shah@redhat.com>
Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>