numa: Use ibm, architecture-vec-5 to detect form 1 affinity
Patchwork-id: 25918
O-Subject: [PATCH RHEL6.0 BZ594502 2/3] powerpc: numa: Use ibm,
	architecture-vec-5 to detect form 1 affinity
Bugzilla: 594502

RHBZ#:
======
https://bugzilla.redhat.com/show_bug.cgi?id=594502

Description:
============

I've been told that the architected way to determine we are in form 1
affinity mode is by reading the ibm,architecture-vec-5 property which
mirrors the layout of the fifth vector of the ibm,client-architecture
structure.

Eventually we may want to parse the ibm,architecture-vec-5 and create
FW_FEATURE_* bits.

Signed-off-by: Anton Blanchard <anton <at> samba.org>