tg3: Provide more support for 57765 with v3.108
Patchwork-id: 25920
O-Subject: [RHEL6 PATCH REPOST] tg3: Provide more support for 57765 with v3.108
Bugzilla: 581691
RH-Acked-by: Andy Gospodarek <gospo@redhat.com>
RH-Acked-by: David S. Miller <davem@redhat.com>
RH-Acked-by: Stanislaw Gruszka <sgruszka@redhat.com>
RH-Acked-by: Ivan Vecera <ivecera@redhat.com>

Repost of the original patch because there was a conflict with
another patch to tg3.c and broadcom.c. Enclosed please find the
updated patch that Aris generated with my assistance. For what
it's worth, I personally checked this with the original and its
conflict patch. When applied it compiled for me without issue.

bz581691
https://bugzilla.redhat.com/bugzilla/show_bug.cgi?id=581691
tg3: Missing 57765 device IDs

Description of problem:
Whereas the device IDs for the 57765 device were missing, there
were a number of significant patches that allowed the device to
work and not panic.

Solution:
This patch will allow for the 57765 to function. In doing so, it will
bring the tg3 driver up to version 3.108.

Upstream status:
All patches can be found upstream.

Acks would be appreciated. Thanks.

Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>