Fall back to GART if initialization fails
Patchwork-id: 25981
O-Subject: [RHEL6 PATCH 2/2] Fall back to GART if initialization fails
Bugzilla: 598974
RH-Acked-by: Prarit Bhargava <prarit@redhat.com>

>From dd0a5301279e542181184c7c5b31d3dc6f614c18 Mon Sep 17 00:00:00 2001
From: root <root@localhost.localdomain>
Date: Wed, 2 Jun 2010 10:04:31 -0400
Subject: [PATCH] RHEL6: Fall back to GART if initialization fails

This patch implements a fallback to the GART IOMMU if this
is possible and the AMD IOMMU initialization failed.
Otherwise the fallback would be nommu which is very
problematic on machines with more than 4GB of memory or
swiotlb which hurts io-performance.

Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>