xen: don't touch xsave in cr4
Patchwork-id: 26022
O-Subject: [RHEL6 PATCH] xen: don't touch xsave in cr4
Bugzilla: 599069
RH-Acked-by: Paolo Bonzini <pbonzini@redhat.com>
RH-Acked-by: Avi Kivity <avi@redhat.com>
RH-Acked-by: Don Dutile <ddutile@redhat.com>

https://bugzilla.redhat.com/show_bug.cgi?id=599069

Legacy hypervisors (RHEL 5.0 and RHEL 5.1) do not handle guest writes to
cr4 gracefully. If a guest attempts to write a bit of cr4 that is
unsupported, then the HV is so offended it crashes the domain. While
later guest kernels (such as RHEL6) don't assume the HV supports all
features, they do expect nicer responses. That assumption introduced
code that probes whether or not xsave is supported early in the boot. So
now when attempting to boot a RHEL6 guest on RHEL5.0 or RHEL5.1 an early
crash will occur.

This patch is quite obviously an undesirable hack. The real fix for this
problem should be in the HV, and is, in later HVs. However, to support
running on old HVs, RHEL6 can take this small change. No impact will
occur for running on any RHEL HV (not even RHEL 5.5 supports xsave).
There is only potential for guest performance loss on upstream Xen.

brew: https://brewweb.devel.redhat.com/taskinfo?taskID=2495385

Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>