nmi_watchdog: disable correct cpu if it fails check
Patchwork-id: 26051
O-Subject: [PATCH 1/2 RHEL-6] [x86_64] nmi_watchdog: disable correct cpu if it
	fails check
Bugzilla: 596760
RH-Acked-by: Prarit Bhargava <prarit@redhat.com>
RH-Acked-by: Dean Nelson <dnelson@redhat.com>
RH-Acked-by: Stefan Assmann <sassmann@redhat.com>

https://bugzilla.redhat.com/show_bug.cgi?id=596760

During boot up, there is a section of code the tries to check to see if
the performance counters are correctly working and generating NMIs.  If
a particular cpu does not generate enough NMIs in a short period, the cpu
is consider broken with regards to its performance counters.  As a result
the nmi watchdog is supposed to be disabled on that cpu.

When I added the hrtimer code, I accidentally canceled the timer on the
wrong cpu and as a result a false LOCKUP was triggered.  This patch fixes
that.

Of course, this is really moot because the system should never have a
broken performance counter subsytem to begin with, which is why no one
really sees this bug (execpt for the bz reporter :-)).

Signed-off-by: Don Zickus <dzickus@redhat.com>
Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>