nmi_watchdog: relax the nmi checks during bootup
Patchwork-id: 26052
O-Subject: [PATCH 2/2 RHEL-6] [x86_64] nmi_watchdog: relax the nmi checks during
	bootup
Bugzilla: 596760
RH-Acked-by: Prarit Bhargava <prarit@redhat.com>
RH-Acked-by: Dean Nelson <dnelson@redhat.com>
RH-Acked-by: Stefan Assmann <sassmann@redhat.com>

https://bugzilla.redhat.com/show_bug.cgi?id=596760

I have been getting a bunch of bug reports that look like this
variety during boot up:

WARNING: CPU#6: NMI appears to be stuck (268->273)!

Recently this has been seen on big boxes, ones with lots of cpus.
Unfortnately, these are false positives as the nmi watchdog is working
fine, it is the way the check is designed that is causing issues.

The way the nmi watchdog check is designed, it sets the performance
counters to trigger every ms.  It then kicks off a smp_call_function
to have every cpu sit in a busy-wait loop to generate cpu cycles that
in turn increment the performance counters to generate NMIs.  After 20ms
the code checks every cpu to see how many NMIs fired.  Now in theory 20
NMIs should have fired (1 for every ms).  The code tries to be
conservative and checks to see at least 5 triggered.  If less than 5
happened, then fail the check.

Now with normal desktops, this logic seems to work great.  Not much for
issues.  Unfortunately with bigger boxes, it seems that lots of stuff
goes on behind the scenes (perhaps other interrupts, SMIs) that starve
the NMI during that period of time.  As a result only a handle of NMIs
are triggered instead of the theoretical 20.

The way the performance counter works, you program a large negative
number and every non-halted clock cycle increments that number.  Once it
crosses zero an NMI is generated.  But the counter keeps incrementing.
So during debugging I checked for large positive values in the counter
when the NMI was finally serviced.   And sure enough, the cpus that
were considered stuck were the same ones with large values in that
counter indicating they were being starved some how.

It would be a complete was of my time to go from machine to machine
to figure out why they are being starved, so instead I changed the
rules for the check to assume if the nmi watchdog can fire at least
once, then it is good to go.

I also increased the time from 20ms to 40ms to give some boxes extra
time (one 48 cpu box always had a dozen or so cpus only triggering 6-7
NMIs in the 20ms window, at 40ms it is up to 12-14).

These are probably not the best solutions, but in reality it is kinda
impossible to control how many interrupts/nmi can be generated in a
small window.  I mean just read the timekeeping code to see the hoops
they go through to see if SMIs were generated during their calibration.

This patch should drastically cut down the number of false positives.

One can argue why have the check at all if it is such a headache.
Well it does catch one class of problems, the unsupported cpu (which is
easily fixed by adding the cpu ids into the table assuming the re-use
the same performance counter logic).

Of course this patch is RHEL-6 only because upstream will have my
rewritten nmi watchdog that sits on top of the perf subsystem.

Signed-off-by: Don Zickus <dzickus@redhat.com>
Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>