fix refcount bug in anon_vma code
Patchwork-id: 26077
O-Subject: [RHEL6 PATCH] fix refcount bug in anon_vma code
Bugzilla: 602739
RH-Acked-by: Larry Woodman <lwoodman@redhat.com>

With the new anon_vma code we take a refcount on the root anon_vma.
However, the root anon_vma does not have a refcount on itself, so
we should not try to do a drop on itself when it is being unlinked.

Just got merged in the -mm kernel.

Fixes bug 602739

Signed-off-by: Rik van Riel <riel@redhat.com>
Tested-by: Dave Young <hidave.darkstar@gmail.com>

Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>