Bluetooth: Keep a copy of each HID device's report descriptor
Patchwork-id: 26084
O-Subject: [RHEL6 5/5] Bluetooth: Keep a copy of each HID device's report
	descriptor
Bugzilla: 565583
RH-Acked-by: Aristeu Rozanski <aris@redhat.com>

backport changeset 15c697ce1c5b408c5e20dcdc6aea2968d1125b75

https://bugzilla.redhat.com/show_bug.cgi?id=565583

The report descriptor is read by user space (via the Service
Discovery Protocol), so it is only available during the ioctl
to connect. However, the HID probe function that needs the
descriptor might not be called until a specific module is
loaded. Keep a copy of the descriptor so it is available for
later use.

Signed-off-by: Mauro Carvalho Chehab <mchehab@redhat.com>
Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>