Remove timeout logic in mutex_spin_on_owner() to match upstream
Patchwork-id: 26091
O-Subject: [PATCH RHEL6.0 BZ602805] Remove timeout logic in
	mutex_spin_on_owner() to match upstream
Bugzilla: 602805
RH-Acked-by: Rik van Riel <riel@redhat.com>
RH-Acked-by: David Howells <dhowells@redhat.com>

RHBZ#:
======
https://bugzilla.redhat.com/show_bug.cgi?id=602805

Description:
============
Currently, we can hit a nasty case with optimistic spinning on
mutexes:

CPU A tries to take a mutex, while holding the BKL

CPU B tried to take the BLK while holding the mutex

This looks like a AB-BA scenario but in practice, is allowed and
happens due to the auto-release-on-schedule nature of the BKL.

In that case, the optimistic spinning code can get us into a situation
where instead of going to sleep, A will spin waiting for B who is
spinning waiting for A, and the only way out of that loop is the
need_resched() test in mutex_spin_on_owner().

The patches committed for 591735 fix both in a rather crude way.
They completely disable spinning if we own the BKL, and they add a safety
timeout using jiffies to fallback to sleeping if we end up spinning for more
than 1 or 2 jiffies. However, the safety timeout was eventually jettisoned
upstream, so this revised patch removes it from RHEL6 as well.

Patch 2 is located here
http://post-office.corp.redhat.com/archives/rhkernel-list/2010-May/msg01145.html

Even though Ingo took Patch 2 into his tree Linus rejected it.  A
discussion as to why is at:
http://lkml.org/lkml/2010/4/28/12

A second version of the patch has been taken by Ingo that would seem to
satisfy Linus.  The commit ID for this second version is:

http://git.kernel.org/?p=linux/kernel/git/x86/linux-2.6-tip.git;a=commitdiff;h=fd6be105b883244127a734ac9f14ae94a022dcc0

Signed-off-by: Tony Breeds <tony@bakeyournoodle.com>
Acked-by: Linus Torvalds <torvalds@linux-foundation.org>
Acked-by: Peter Zijlstra <a.p.zijlstra@chello.nl>
Cc: Benjamin Herrenschmidt <benh@kernel.crashing.org>
Cc: <stable@kernel.org>
LKML-Reference: <20100519054636.GC12389@ozlabs.org>
[ added an unlikely() attribute to the branch ]
Signed-off-by: Ingo Molnar <mingo@elte.hu>

RHEL Version Found:
===================
RHEL 6.0

kABI Status:
============
No symbols were harmed.

Brew:
=====
http://brewweb.devel.redhat.com/brew/taskinfo?taskID=2509511

Test Status:
============
Tested by the IBM test team.

===============================================================
Steve Best
IBM on-site partner

Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>