Fix a hang of khubd if UHCI is removed
Patchwork-id: 26098
O-Subject: [RHEL 6 patch] Fix a hang of khubd if UHCI is removed
Bugzilla: 579093
RH-Acked-by: Rik van Riel <riel@redhat.com>
RH-Acked-by: Don Zickus <dzickus@redhat.com>

The bz#579093 is for the hang in UHCI in testing at Stratus.
This is different from the crash in HID which I fixed recently,
although triggered by the same test procedure.

The backtrace is wholly uninteresting, it's always the same usb_kill_urb
that occurs when the Host Controller (HC) fails to deliver an interrupt.

The fix is rather involved, unfortunately. The important part is to
fake interrupts when we know that real ones aren't coming. That
unblocks usb_kill_urb. However, it's not all and if we do that
things will crash, so the rest of the patch does various cleanups
to the way the pesky timers are stopped.

The customer did the development together with the upstream maintainer
and tested various iterations of this, including the final binary
kernel which I build in Brew. So we know this works. It also basically
works for common computers.

I consider risk of this low. It is not zero (it touches sysfs after
all :-), but looks acceptable to me. It should not affect anything
until we start removing hardware.

The patch is not in upstream, but since it was actually written
by Alan Stern, I do not expect any friction getting it in.

Please ack.

-- Pete

commit 78038d936e3bf2ea2cbc99ee406b294034cc884e
Author: Master <zaitcev@lembas.zaitcev.lan>
Date:   Tue Jun 8 16:58:01 2010 -0600

    khubd hang on removal of UHCI device (bz579093) -- Alan 2010/6/4

Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>