always invalidate and flush on spte page size change
Patchwork-id: 26102
O-Subject: [PATCH 09 of 10 RHEL6] always invalidate and flush on spte page size
	change
Bugzilla: 578134
RH-Acked-by: Rik van Riel <riel@redhat.com>
RH-Acked-by: Larry Woodman <lwoodman@redhat.com>

From: Marcelo Tosatti <mtosatti@redhat.com>

Always invalidate spte and flush TLBs when changing page size, to make
sure different sized translations for the same address are never cached
in a CPU's TLB.

The first case where this occurs is when a non-leaf spte pointer is
overwritten by a leaf, large spte entry. This can happen after dirty
logging is disabled on a memslot, for example.

The second case is a leaf, large spte entry is overwritten with a
non-leaf spte pointer, in __direct_map. Note this cannot happen now
because the only potential source of such overwrite is dirty logging
being enabled, which zaps all MMU pages. But this might change
in the future, so better be robust against it.

Noticed by Andrea.

KVM-Stable-Tag

Signed-off-by: Marcelo Tosatti <mtosatti@redhat.com>