cifs: don't attempt busy-file rename unless it's in same directory
Patchwork-id: 26113
O-Subject: [RHEL6 PATCH] BZ#603707: cifs: don't attempt busy-file rename unless
	it's in same directory
Bugzilla: 603707
RH-Acked-by: Steve Dickson <SteveD@redhat.com>

Older windows SMB servers typically refuse to allow files that are open
to be renamed as the underlying filesystem doesn't support it. If you
rename by filehandle however, you can generally rename files as long as
that doesn't cross directories.

The CIFS code is set up to try to rename by filehandle when renaming
by name fails, but it doesn't properly limit that code to renames
within the same directory. When that is attempted, the file is generally
renamed to the target name, but stays within the same directory.

The following patch is being pushed upstream for 2.6.35 and into stable
trees. It just prevents the busy-file rename code from being attempted
when the source and target have different parent directories.

----------------------[snip]---------------------

Busy-file renames don't actually work across directories, so we need
to limit this code to renames within the same dir.

This fixes the bug detailed here:

    https://bugzilla.redhat.com/show_bug.cgi?id=591938

Signed-off-by: Jeff Layton <jlayton@redhat.com>
CC: Stable <stable@kernel.org>
Signed-off-by: Steve French <sfrench@us.ibm.com>
Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>