Move kdump default base address to 64MB on 64bit
Patchwork-id: 26184
O-Subject: [PATCH RHEL6.0 BZ603779] powerpc: Move kdump default base address to
	64MB on 64bit
Bugzilla: 603779
RH-Acked-by: David Howells <dhowells@redhat.com>

RHBZ#:
======
https://bugzilla.redhat.com/show_bug.cgi?id=603779

Description:
============
We are seeing boot fails on some System p machines when using the kdump
crashkernel= boot option. The default kdump base address is 32MB, so if we
reserve 256MB for kdump then we reserve all of the RMO except the first 32MB.

We really want kdump to reserve some memory in the RMO and most of it
elsewhere but that will require more significant changes. For now we can shift
the default base address to 64MB when CONFIG_PPC64 and CONFIG_RELOCATABLE are
set. This isn't quite correct since what we really care about is the kdump
kernel is relocatable, but we already make the assumption that base kernel
and kdump kernel have the same CONFIG_RELOCATABLE setting, eg:

#ifndef CONFIG_RELOCATABLE
        if (crashk_res.start != KDUMP_KERNELBASE)
                printk("Crash kernel location must be 0x%x\n",
                                KDUMP_KERNELBASE);
...

RTAS is instantiated towards the top of our RMO, so if we were to go any
higher we risk not having enough RMO memory for the kdump kernel on boxes
with a 128MB RMO.

Signed-off-by: Anton Blanchard <anton@samba.org>

RHEL Version Found:
===================
RHEL 6.0

kABI Status:
============
No symbols were harmed.

Brew:
=====
https://brewweb.devel.redhat.com/taskinfo?taskID=2515703

Upstream Status:
================

in the process of being committed upstream.

patch is here

http://patchwork.ozlabs.org/patch/54923/

Test Status:
============
Tested by the IBM test team. Not all systems will see this issue.
Tested on the system that had the issue and the patch fixed it and
this system now boots.

===============================================================
Steve Best
IBM on-site partner

Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>