fat: fix buffer overflow in vfat_create_shortname()
Patchwork-id: 26187
O-Subject: [RHEL6 PATCH 1/63] fat: fix buffer overflow in vfat_create_shortname()
Bugzilla: 604608
RH-Acked-by: Rik van Riel <riel@redhat.com>
RH-Acked-by: Eric Sandeen <sandeen@redhat.com>

From: Nikolaus Schulz <microschulz@web.de>

commit 30d1872d9eb3663b4cf7bdebcbf5cd465674cced upstream.

When using the string representation of a random counter as part of the base
name, ensure that it is no longer than 4 bytes.

Since we are repeatedly decrementing the counter in a loop until we have found a
unique base name, the counter may wrap around zero; therefore, it is not enough
to mask its higher bits before entering the loop, this must be done inside the
loop.

[hirofumi@mail.parknet.co.jp: use snprintf()]
Signed-off-by: Nikolaus Schulz <microschulz@web.de>
Signed-off-by: OGAWA Hirofumi <hirofumi@mail.parknet.co.jp>
Signed-off-by: Linus Torvalds <torvalds@linux-foundation.org>
Signed-off-by: Greg Kroah-Hartman <gregkh@suse.de>
Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>