cifs: move cifs_new_fileinfo call out of cifs_posix_open
Patchwork-id: 26303
O-Subject: [RHEL6 PATCH 3/7] BZ#593422: cifs: move cifs_new_fileinfo call out of
	cifs_posix_open
Bugzilla: 593422
RH-Acked-by: Steve Dickson <SteveD@redhat.com>

Having cifs_posix_open call cifs_new_fileinfo is problematic and
inconsistent with how "regular" opens work. It's also buggy as
cifs_reopen_file calls this function on a reconnect, which creates a new
struct cifsFileInfo that just gets leaked.

Push it out into the callers. This also allows us to get rid of the
"mnt" arg to cifs_posix_open.

Finally, in the event that a cifsFileInfo isn't or can't be created, we
always want to close the filehandle out on the server as the client
won't have a record of the filehandle and can't actually use it. Make
sure that CIFSSMBClose is called in those cases.

Signed-off-by: Jeff Layton <jlayton@redhat.com>
Reviewed-and-Tested-by: Suresh Jayaraman <sjayaraman@suse.de>

Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>