virtio: net: Remove net-specific advertising of PUBLISH_USED feature
Patchwork-id: 26306
O-Subject: [RHEL6 PATCH] virtio: net: Remove net-specific advertising of
	PUBLISH_USED feature
Bugzilla: 605591
RH-Acked-by: Jes Sorensen <Jes.Sorensen@redhat.com>
RH-Acked-by: Juan Quintela <quintela@redhat.com>

Bugzilla: 605591

Commit fb03d0d8b9d2006a6b7387d6ec9822bf3d3391af added a new feature to
the virtio ring infrastructure. This feature was meant to be exposed
only for virtio-net at the time. However, it inadvertently got
advertised for all virtio devices.

The reason to limit to virtio-net was because some perf regression was
seen on virtio-blk tests. Jes has run tests again and has found about a
20% reduction in vmexit rates, w/o impacting the performance. No perf
regression was found for virtio-console as well.

So it's now OK to enable this feature for all devices. Bug 602417 tracks
the qemu side of things.

Since the bug mentioned above exposed this feature for all virtio
devices anyway, just remove the virtio-net specific code for exposing
this feature.

[ The hunk

x@ -471,6 +472,8 @@ void vring_transport_features(struct virtio_device
*vdev)
                switch (i) {
                case VIRTIO_RING_F_INDIRECT_DESC:
                        break;
+               case VIRTIO_RING_F_PUBLISH_USED:
+                       break;
                default:
                        /* We don't understand this bit. */
                        clear_bit(i, vdev->features);

in the commit referenced above enabled this feature for all the devices]

Signed-off-by: Amit Shah <amit.shah@redhat.com>
Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>