spec: modify %post script to not overwrite crashkernel line
Patchwork-id: 26339
O-Subject: [RHEL6 PATCH] spec: modify %post script to not overwrite crashkernel
	line (bz 596919)
Bugzilla: 596919
RH-Acked-by: Stanislaw Gruszka <sgruszka@redhat.com>
RH-Acked-by: Aristeu Rozanski <aris@redhat.com>
RH-Acked-by: Prarit Bhargava <prarit@redhat.com>
RH-Acked-by: Don Zickus <dzickus@redhat.com>

Hey all-
	Currently the installation of a new kernel overwrites the crashkernel
parameter of the existing default kernel.  This is bad if a customer has
overridden that parameter with their own setting.  This patch detects if a
crashkernel parameter already exists and leaves it unmodified in the event that
it does.  Tested successfully by me, fixes bz 596919.

Neil

Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>