enable kmemleak for the -debug kernel
Patchwork-id: 26366
O-Subject: [RHEL6 PATCH] enable kmemleak for the -debug kernel
Bugzilla: 252066
RH-Acked-by: Rik van Riel <riel@redhat.com>
RH-Acked-by: Prarit Bhargava <prarit@redhat.com>

Hi,

There is a customer request to enabled 'kmemleak'. However, due to its
overhead we can only turn it on in our -debug variant.

Currently, there is a Kconfig restriction on kmemleak that requries
!MEMORY_HOTPLUG, and we have MEMORY_HOTPLUG set for both our stock
and -debug variants. I have proposed an upstream patch to resolve the
issue, but thus far, it has not been picked up:

http://marc.info/?l=linux-kernel&m=126933638526171&w=2

That said, I don't think we need to wait for this patch to be included
upstream before setting kmemleak in our -debug kernel variant. First
off, testing memory hotplug while running kmemleak is going to be a
pretty rare occurence. And even if they are run together, I have
successfully tested them several times (did not run into the race
conditions).

There is also a 'CONFIG_DEBUG_KMEMLEAK_EARLY_LOG_SIZE' parameter that
needs to be set with this patch. In my testing on a 64GB system, I
needed around 5,000 entries to successully boot. Thus, I am proposing a
20,000 value for this config param, which has a range of 200 - 40,000 with
a default of 400. Basically, this parameter reserves, a number of
records in __initdata. The size of each record is 168 bytes. Thus, with
a value of 20,000, this is less than 3 MB, which are subsequently
freed.

Thus, I am proposing the patch below, resolves bz #252066.
Also, successfully brewed.

thanks,

-Jason

Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>