ehca: bitmask handling for lock_hcalls
Patchwork-id: 26430
O-Subject: [PATCH RHEL6.0 BZ605739] ib/ehca: bitmask handling for lock_hcalls
Bugzilla: 605739
RH-Acked-by: David Howells <dhowells@redhat.com>

RHBZ#:
======
https://bugzilla.redhat.com/show_bug.cgi?id=605739

Description:
============
Fix reading capability bit from device capabilities for hcall locking.

Signed-off-by: Alexander Schmidt <alexs@linux.vnet.ibm.com>

ehca hcall bitmask patch

When parsing the device capabilities, an invalid bitmask was used to check if
hcall locking should be enabled or not.

Without this fix, hcall locking might be enabled on systems where it should not
be enabled. This wrong setting may cause severe performance issues in clustered
environments up to complete inresponsiveness of single nodes due to locking contention.

RHEL Version Found:
===================
RHEL 6.0

kABI Status:
============
No symbols were harmed.

Brew:
=====
http://brewweb.devel.redhat.com/brew/taskinfo?taskID=2534356

Upstream status:
===============
patch posted here
http://marc.info/?l=linux-kernel&m=127616083201314&w=2

I talked to Alexander about the upstream status of this one and he provided
the following info.

This patch was posted to the mailing list, and we got asked from the IB maintainer
if we want to get this included in the current development cycle (2.6.35)
or the next one (2.6.36). We decided that the patch should
go into 2.6.36, so there is no git commit yet, as the development of 2.6.36
has not started yet.

Link to comments from IB maintainer:
http://marc.info/?l=linux-rdma&m=127627217124190&w=2

Test Status:
============
This patch was applied and tested on kernel version 2.6.32-33.el6.ppc64.
It was verified that this patch fixes the issue.

===============================================================
Steve Best
IBM on-site partner

Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>