vt/fbcon: try harder to print output when panicing
Patchwork-id: 26450
O-Subject: [PATCH] [RHEL6.0 PATCH] vt/fbcon/drm: try harder to print output when
	panicing
Bugzilla: 579002
RH-Acked-by: Adam Jackson <ajax@redhat.com>

bz 579002

upstream status: in drm-next now, scheduled for 2.6.36, posted to lkml.

this rolls up 3 upstream patches:
fb: fix colliding defines for fb flags.
drm: add locked variant of drm_fb_helper_force_kernel_mode
vt/console: try harder to print output when panicing

The 3rd patch is the main one, and it allows for fbcon users to set a flag
that means that we will try and force write to the console even if X is
running, this means when a panic finishes we can force the output to that
console and see the panic outputs.

Tested used lkdtm while X is running.

Signed-off-by: Dave Airlie <airlied@redhat.com>
Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>