swap: Add swap slot free callback to block_device_operations
Patchwork-id: 26457
O-Subject: [RHEL6 PATCH 3/9] swap: Add swap slot free callback to
	block_device_operations
Bugzilla: 578641
RH-Acked-by: Rik van Riel <riel@redhat.com>

Bugzilla:
https://bugzilla.redhat.com/show_bug.cgi?id=578641

This callback is required when RAM based devices are used as swap disks.
One such device is ramzswap which is used as compressed in-memory swap
disk.  For such devices, we need a callback as soon as a swap slot is no
longer used to allow freeing memory allocated for this slot.  Without this
callback, stale data can quickly accumulate in memory defeating the whole
purpose of such devices.

Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>