GFS2: recovery stuck on transaction lock
Patchwork-id: 26475
O-Subject: [RHEL6 Patch] bz#590878 - GFS2: recovery stuck on transaction lock
Bugzilla: 590878
RH-Acked-by: Steven Whitehouse <swhiteho@redhat.com>

Hi,

This patch fixes bugzilla bug #590878: GFS2: recovery stuck on
transaction lock.  We set the frozen flag on the glock when we receive
a completion that cannot be delivered due to blocked locks. At that
point we check to see whether the first waiting holder has the noexp
flag set. If the noexp lock is queued later, then we need to unfreeze
the glock at that point in time, namely, in the glock work function.

This patch was originally written by Steve Whitehouse, but since
he's on holiday, I'm submitting it.  I've submitted it upstream as
well, but it hasn't been accepted yet due to Steve's vacation.
It has been well tested with a complex recovery test called revolver.

Regards,

Bob Peterson
Red Hat GFS

Signed-off-by: Steve Whitehouse <swhiteho@redhat.com>
Signed-off-by: Bob Peterson <rpeterso@redhat.com>
--

Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>